+++ include/mxnet/ndarray.h
@@ -594,7 +594,7 @@ class NDArray {
                    std::vector<std::string>* keys);
 
  private:
-  friend class ImperativeRuntime;
+  friend class Imperative;
   /*! \brief the real data chunk that backs NDArray */
   // shandle is used to store the actual values in the NDArray
   // aux_handles store the aux data(such as indices) if it's needed by non-default storage.
