+++ src/imperative/imperative.cc
@@ -1,0 +1,575 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+#include <unordered_set>
+#include <iostream>
+#include "./utils.h"
+
+namespace mxnet {
+#if DMLC_CXX11_THREAD_LOCAL
+thread_local bool Imperative::is_train_ = false;
+thread_local bool Imperative::is_recording_ = false;
+#else
+MX_THREAD_LOCAL bool Imperative::is_train_ = false;
+MX_THREAD_LOCAL bool Imperative::is_recording_ = false;
+#endif
+
+Imperative* Imperative::Get() {
+  static Imperative inst;
+  return &inst;
+}
+
+OpStatePtr Imperative::InvokeOp(
+    const Context& ctx,
+    const nnvm::NodeAttrs& attrs,
+    const std::vector<NDArray*>& inputs,
+    const std::vector<NDArray*>& outputs,
+    const std::vector<OpReqType>& req,
+    OpStatePtr state) {
+  using namespace imperative;
+  static auto& createop = nnvm::Op::GetAttr<FCreateOpState>("FCreateOpState");
+  static auto& is_layer_backward = Op::GetAttr<bool>("TIsLayerOpBackward");
+  MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
+
+  const nnvm::Op *op = attrs.op;
+
+  std::vector<engine::VarHandle> read_vars, write_vars;
+  std::vector<Resource> requested;
+  std::vector<uint32_t> mutate_idx;
+  SetDependency(attrs, ctx, inputs, outputs,
+      &read_vars, &write_vars, &requested, &mutate_idx);
+
+  FCompute fn = common::GetFCompute<FCompute>(op, "FCompute", ctx);
+  FComputeEx fn_ex = common::GetFCompute<FComputeEx>(op, "FComputeEx", ctx);
+
+  bool has_non_default = common::ContainsNonDefaultStorage(inputs) ||
+                         common::ContainsNonDefaultStorage(outputs);
+  if (fn && (!has_non_default || !fn_ex)) {
+    PushFCompute(fn, op, attrs, ctx, read_vars, write_vars,
+        requested, inputs, outputs, mutate_idx, req);
+  } else if (fn_ex) {
+    PushFComputeEx(fn_ex, op, attrs, ctx, read_vars, write_vars,
+        requested, inputs, outputs, req);
+  } else if (createop.count(op) || is_layer_backward.get(op, false)) {
+    if (!state) {
+      state = createop[op](attrs, ctx, ret->arg_shapes, ret->arg_types);
+    }
+    write_vars.push_back(state.get_var());
+    PushOperator(state, op, attrs, ctx, read_vars, write_vars,
+        requested, inputs, outputs, mutate_idx, req);
+  } else {
+    LOG(FATAL)
+      << "Operator " << op->name << " is not implemented for "
+      << (ctx.dev_mask() == gpu::kDevMask ? "GPU." : "CPU.");
+  }
+
+  return state;
+}
+
+OpStatePtr Imperative::Invoke(
+    const Context& default_ctx,
+    const nnvm::NodeAttrs& attrs,
+    const std::vector<NDArray*>& inputs,
+    const std::vector<NDArray*>& outputs) {
+  using namespace imperative;
+  static auto& ndfunc = nnvm::Op::GetAttr<FNDArrayFunction>("FNDArrayFunction");
+
+  if (ndfunc.count(attrs.op)) {
+    std::vector<NDArray> p_inputs, p_outputs;
+    DerefInputOutput(inputs, outputs, &p_inputs, &p_outputs);
+    ndfunc[attrs.op](attrs, p_inputs, &p_outputs);
+    for (size_t i = 0; i < outputs.size(); ++i) *outputs[i] = std::move(p_outputs[i]);
+    return OpStatePtr();
+  }
+
+  // TODO(piiswrong): infer ctx
+  Context ctx = GetContext(attrs, inputs, outputs, default_ctx);
+  SetShapeType(ctx, attrs, inputs, outputs);
+  std::vector<OpReqType> req;
+  SetWriteInplaceReq(inputs, outputs, &req);
+
+  return InvokeOp(ctx, attrs, inputs, outputs, req);
+}
+
+void Imperative::MarkVariables(
+    const std::vector<NDArray*>& variables,
+    const std::vector<mx_uint>& grad_reqs,
+    const std::vector<NDArray*>& gradients) {
+  for (uint32_t i = 0; i < variables.size(); ++i) {
+    std::string str_c(std::to_string(variable_count_++));
+
+    variables[i]->entry_ = nnvm::NodeEntry{
+        nnvm::Symbol::CreateVariable("var" + str_c).outputs[0].node, 0, 0};
+    AGInfo& info = AGInfo::Create(variables[i]->entry_.node);
+    info.outputs.emplace_back(variables[i]->Detach());
+    info.out_grads.emplace_back(gradients[i]->Detach());
+    info.grad_req = static_cast<OpReqType>(grad_reqs[i]);
+
+    gradients[i]->entry_ = nnvm::NodeEntry{
+        nnvm::Symbol::CreateVariable("grad" + str_c).outputs[0].node, 0, 0};
+    AGInfo& grad_info = AGInfo::Create(gradients[i]->entry_.node);
+    grad_info.outputs.emplace_back(gradients[i]->Detach());
+  }
+}
+
+
+void Imperative::GetBackwardDependency(
+    const nnvm::NodePtr& node,
+    uint32_t num_inputs, uint32_t num_outputs,
+    std::vector<bool> *p_save_inputs,
+    std::vector<bool> *p_save_outputs) {
+  static auto& fgradient = nnvm::Op::GetAttr<nnvm::FGradient>("FGradient");
+  std::vector<bool>& save_inputs = *p_save_inputs;
+  std::vector<bool>& save_outputs = *p_save_outputs;
+  save_inputs.resize(num_inputs);
+  save_outputs.resize(num_outputs);
+  std::fill(save_inputs.begin(), save_inputs.end(), false);
+  std::fill(save_outputs.begin(), save_outputs.end(), false);
+
+  node->inputs.clear();
+  node->inputs.reserve(num_inputs);
+  for (uint32_t i = 0; i < num_inputs; ++i) {
+    node->inputs.emplace_back(nnvm::NodeEntry{nullptr, i, 0});
+  }
+
+  if (fgradient.count(node->op())) {
+    std::vector<nnvm::NodeEntry> ograd_entries;
+    ograd_entries.reserve(num_outputs);
+    for (uint32_t i = 0; i < num_outputs; ++i) {
+      ograd_entries.emplace_back(nnvm::NodeEntry{nullptr, i, 1});
+    }
+    auto igrad_entries = fgradient[node->op()](node, ograd_entries);
+    for (const auto& i : igrad_entries) {
+      if (i.node == nullptr && i.version == 0) {
+        save_inputs[i.index] = true;
+      } else if (i.node == node) {
+        save_outputs[i.index] = true;
+      }
+    }
+    DFSVisit(igrad_entries, [&](const nnvm::NodePtr& gnode) {
+        if (!gnode || gnode == node) return;
+        for (const auto& i : gnode->inputs) {
+          if (i.node == nullptr && i.version == 0) {
+            save_inputs[i.index] = true;
+          } else if (i.node == node) {
+            save_outputs[i.index] = true;
+          }
+        }
+      });
+  }
+}
+
+void Imperative::RecordOp(
+    nnvm::NodeAttrs&& attrs,
+    const std::vector<NDArray*>& inputs,
+    const std::vector<NDArray*>& outputs,
+    const OpStatePtr& state,
+    std::vector<bool>* p_save_inputs,
+    std::vector<bool>* p_save_outputs) {
+  MXAPIThreadLocalEntry *local_buff = MXAPIThreadLocalStore::Get();
+
+  for (uint32_t i = 0; i < outputs.size(); ++i) {
+    CHECK(AGInfo::IsNone(*(outputs[i])))
+      << "Assigning to NDArrays that are already in a computational graph "
+      << "will cause undefined behavior when evaluating gradients. "
+      << "Please call backward first to clear the graph or do this out side of "
+      << "a record section. ";
+  }
+
+  bool need_grad = false;
+  for (const auto& i : inputs) {
+    if (AGInfo::IsNone(*i)) continue;
+    need_grad = true;
+    break;
+  }
+  if (!need_grad) return;
+
+  nnvm::NodePtr node = nnvm::Node::Create();
+  node->attrs = std::move(attrs);
+  node->attrs.name = "node_" + std::to_string(node_count_++);
+  AGInfo& info = AGInfo::Create(node);
+  info.state = state;
+
+  if (p_save_inputs == nullptr) {
+    p_save_inputs = &(local_buff->save_inputs);
+    p_save_outputs = &(local_buff->save_outputs);
+    GetBackwardDependency(
+        node, inputs.size(), outputs.size(), p_save_inputs, p_save_outputs);
+  } else {
+    node->inputs.resize(inputs.size());
+  }
+
+  std::vector<bool>& save_inputs = *p_save_inputs;
+  std::vector<bool>& save_outputs = *p_save_outputs;
+
+  for (size_t i = 0; i < inputs.size(); ++i) {
+    if (AGInfo::IsNone(*(inputs[i]))) {
+      nnvm::NodeEntry entry{nnvm::Symbol::CreateVariable(
+          "null" + std::to_string(variable_count_++)).outputs[0].node, 0, 0};
+      AGInfo& input_info = AGInfo::Create(entry.node);
+      if (save_inputs[i]) {
+        input_info.outputs.emplace_back(*inputs[i]);
+      } else {
+        // Put a dummy array here since it will not be used.
+        input_info.outputs.emplace_back();
+        input_info.outputs.back().shape_ = inputs[i]->shape();
+        input_info.outputs.back().dtype_ = inputs[i]->dtype();
+        input_info.outputs.back().storage_type_ = inputs[i]->storage_type();
+      }
+      inputs[i]->entry_ = std::move(entry);  // assign last to prevent cyclic reference
+    } else if (save_inputs[i]) {
+      AGInfo::Get(inputs[i]->entry_.node).outputs[inputs[i]->entry_.index] = inputs[i]->Detach();
+    }
+    node->inputs[i] = inputs[i]->entry_;
+  }
+
+  for (uint32_t i = 0; i < outputs.size(); ++i) {
+    CHECK(AGInfo::IsNone(*(outputs[i])))
+      << "Inplace operations (+=, -=, x[:]=, etc) are not supported when "
+      << "recording with autograd.";
+  }
+
+  for (uint32_t i = 0; i < outputs.size(); ++i) {
+    if (save_outputs[i]) {
+      info.outputs.emplace_back(outputs[i]->Detach());
+    } else {
+      // Put a dummy array here since it will not be used.
+      info.outputs.emplace_back();
+      info.outputs.back().shape_ = outputs[i]->shape();
+      info.outputs.back().dtype_ = outputs[i]->dtype();
+      info.outputs.back().storage_type_ = outputs[i]->storage_type();
+    }
+    outputs[i]->entry_ = nnvm::NodeEntry{node, i, 0};
+  }
+}
+
+void Imperative::RunGraph(
+    const bool retain_graph,
+    const Context& default_ctx,
+    const nnvm::IndexedGraph& idx,
+    const std::vector<NDArray*> arrays,
+    size_t node_start, size_t node_end,
+    std::vector<OpReqType>&& array_reqs,
+    std::vector<uint32_t>&& ref_count,
+    std::vector<OpStatePtr> *p_states) {
+  using namespace nnvm;
+  using namespace imperative;
+  static auto& createop = nnvm::Op::GetAttr<FCreateOpState>("FCreateOpState");
+  static auto& is_layer_backward = Op::GetAttr<bool>("TIsLayerOpBackward");
+  static const auto bwd_cached_op = Op::Get("_backward_CachedOp");
+
+  std::vector<OpStatePtr>& states = *p_states;
+  bool recording = is_recording();
+
+  std::vector<NDArray*> ndinputs, ndoutputs;
+  ShapeVector arg_shapes;
+  DTypeVector arg_dtypes;
+  std::vector<OpReqType> req;
+
+  for (size_t i = node_start; i < node_end; ++i) {
+    const nnvm::IndexedGraph::Node& node = idx[i];
+    if (node.source->op() == nullptr) continue;
+    ndinputs.clear();
+    ndinputs.reserve(node.inputs.size());
+    for (const auto& j : node.inputs) {
+      ndinputs.emplace_back(arrays[idx.entry_id(j)]);
+      CHECK(!ndinputs.back()->is_none()) << idx[j.node_id].source->attrs.name << " " << j.index;
+    }
+    ndoutputs.clear();
+    ndoutputs.reserve(node.source->num_outputs());
+    req.clear();
+    req.reserve(node.source->num_outputs());
+    for (size_t j = 0; j < node.source->num_outputs(); ++j) {
+      size_t eid = idx.entry_id(i, j);
+      ndoutputs.emplace_back(arrays[eid]);
+      req.push_back(array_reqs[eid]);
+      CHECK(!ndoutputs.back()->is_none());
+    }
+
+    if (node.source->op() == bwd_cached_op) {
+      const auto& cached_op = dmlc::get<CachedOpPtr>(node.source->attrs.parsed);
+      nnvm::Node* fwd_node = node.source->control_deps[0].get();
+      auto fwd_node_id = idx.node_id(fwd_node);
+      cached_op->Backward(retain_graph, states[fwd_node_id], ndinputs, req, ndoutputs);
+    } else if (createop.count(node.source->op())) {
+      arg_shapes.clear();
+      arg_dtypes.clear();
+      arg_shapes.reserve(ndinputs.size());
+      arg_dtypes.reserve(ndinputs.size());
+      for (size_t i = 0; i < ndinputs.size(); ++i) {
+        arg_shapes.emplace_back(ndinputs[i]->shape());
+        arg_dtypes.emplace_back(ndinputs[i]->dtype());
+      }
+      states[i] = createop[node.source->op()](
+          node.source->attrs, default_ctx, arg_shapes, arg_dtypes);
+      InvokeOp(default_ctx, node.source->attrs, ndinputs, ndoutputs, req, states[i]);
+      if (recording) RecordOp(NodeAttrs(node.source->attrs), ndinputs, ndoutputs, states[i]);
+    } else if (is_layer_backward.get(node.source->op(), false)) {
+      nnvm::Node* fwd_node = node.source->control_deps[0].get();
+      auto fwd_node_id = idx.node_id(fwd_node);
+      InvokeOp(default_ctx, node.source->attrs, ndinputs, ndoutputs, req, states[fwd_node_id]);
+      if (recording) {
+        RecordOp(NodeAttrs(node.source->attrs), ndinputs, ndoutputs, states[fwd_node_id]);
+      }
+    } else {
+      InvokeOp(default_ctx, node.source->attrs, ndinputs, ndoutputs, req);
+      if (recording) RecordOp(NodeAttrs(node.source->attrs), ndinputs, ndoutputs);
+    }
+
+    for (const auto& j : node.inputs) {
+      size_t eid = idx.entry_id(j);
+      --ref_count[eid];
+      if (ref_count[eid] == 0) arrays[eid]->ptr_.reset();
+    }
+    for (size_t j = 0; j < ndoutputs.size(); ++j) {
+      size_t eid = idx.entry_id(i, j);
+      if (ref_count[eid] == 0) arrays[eid]->ptr_.reset();
+    }
+  }
+}
+
+
+std::vector<NDArray*> Imperative::Backward(
+    const std::vector<NDArray*>& outputs,
+    const std::vector<NDArray*>& ograds,
+    const std::vector<NDArray*>& variables,
+    bool is_train, bool retain_graph,
+    bool create_graph) {
+  using namespace nnvm;
+  using namespace imperative;
+  static const std::vector<const Op*> zero_ops{Op::Get("zeros_like"), Op::Get("_zeros")};
+
+  // Construct forward graph
+  Graph graph;
+  graph.outputs.reserve(outputs.size());
+  for (const auto& i : outputs) {
+    CHECK(!AGInfo::IsNone(*i))
+      << "Cannot differentiate node because it is not in a computational graph. "
+      << "You need to set is_recording to true or use autograd.record() to save "
+      << "computational graphs for backward. If you want to differentiate the same "
+      << "graph twice, you need to pass retain_graph=True to backward.";
+    graph.outputs.emplace_back(i->entry_);
+  }
+  size_t num_forward_outputs = graph.outputs.size();
+
+  // Prepare head gradients
+  std::vector<NodeEntry> ograd_entries;
+  ograd_entries.reserve(ograds.size());
+  for (size_t i = 0; i < outputs.size(); ++i) {
+    ograd_entries.emplace_back(NodeEntry{Node::Create(), 0, 0});
+    AGInfo& info = AGInfo::Create(ograd_entries.back().node);
+    if (ograds[i] != nullptr) {
+      info.outputs.emplace_back(*ograds[i]);
+    } else {
+      info.outputs.emplace_back(outputs[i]->shape(), outputs[i]->ctx(),
+                                true, outputs[i]->dtype());
+      info.outputs.back() = static_cast<real_t>(1.0);
+    }
+  }
+
+  // Get gradient graph
+  Symbol sym;
+  sym.outputs = graph.outputs;
+  std::vector<NodeEntry> xs;
+  std::vector<NDArray*> x_grads;
+  std::vector<OpReqType> x_reqs;
+  if (variables.size()) {
+    xs.reserve(variables.size());
+    x_grads.reserve(variables.size());
+    x_reqs.reserve(variables.size());
+    for (size_t i = 0; i < variables.size(); ++i) {
+      CHECK(!AGInfo::IsNone(*variables[i]) &&
+            AGInfo::IsVariable(variables[i]->entry_.node))
+          << "Cannot differentiate with respect to the " << i+1 << "-th variable"
+          << " because it does not require gradient.";
+      xs.emplace_back(variables[i]->entry_);
+      x_grads.push_back(new NDArray());
+      x_reqs.push_back(kWriteTo);
+    }
+  } else {
+    std::vector<NodePtr> args = sym.ListInputs(Symbol::kReadOnlyArgs);
+    xs.reserve(args.size());
+    x_grads.reserve(args.size());
+    x_reqs.reserve(args.size());
+    for (const auto& i : args) {
+      AGInfo& info = AGInfo::Get(i);
+      if (info.grad_req == kNullOp) continue;
+      xs.emplace_back(NodeEntry{i, 0, 0});
+      x_grads.push_back(&info.out_grads[0]);
+      x_reqs.push_back(info.grad_req);
+      info.fresh_out_grad = true;
+    }
+    CHECK_GT(xs.size(), 0)
+        << "There are no inputs in computation graph that require gradients.";
+  }
+
+  Graph g_graph = pass::Gradient(
+      graph, graph.outputs, xs, ograd_entries,
+      exec::AggregateGradient, false, nullptr,
+      zero_ops, "_copy");
+  CHECK_EQ(g_graph.outputs.size(), xs.size());
+  for (const auto &e : g_graph.outputs) {
+    graph.outputs.push_back(e);
+  }
+  const auto& idx = graph.indexed_graph();
+  // get number of nodes used in forward pass
+  size_t num_forward_nodes = 0;
+  size_t num_forward_entries = 0;
+  for (size_t i = 0; i < num_forward_outputs; ++i) {
+    num_forward_nodes = std::max(
+        num_forward_nodes, static_cast<size_t>(idx.outputs()[i].node_id + 1));
+    num_forward_entries = std::max(
+        num_forward_entries, static_cast<size_t>(idx.entry_id(idx.outputs()[i])) + 1);
+  }
+
+  // Allocate buffer
+  std::vector<NDArray> buff(idx.num_node_entries());
+  std::vector<uint32_t> ref_count(buff.size(), 0);
+  std::vector<OpStatePtr> states;
+  std::vector<NDArray*> arrays;
+  arrays.reserve(buff.size());
+  for (size_t i = 0; i < buff.size(); ++i) arrays.push_back(&buff[i]);
+  if (create_graph) {
+    states.resize(num_forward_nodes);
+    nnvm::DFSVisit(sym.outputs, [&](const nnvm::NodePtr& n) {
+      AGInfo& info = AGInfo::Get(n);
+      states[idx.node_id(n.get())] = info.state;
+      for (uint32_t i = 0; i < info.outputs.size(); ++i) {
+        CHECK(idx.exist(n.get()));
+        size_t nid = idx.node_id(n.get());
+        size_t eid = idx.entry_id(nid, i);
+        buff[eid] = info.outputs[i];
+        buff[eid].entry_ = NodeEntry{n, i, 0};
+        ref_count[eid] = 1;
+      }
+    });
+    for (size_t i = 0; i < ograd_entries.size(); ++i) {
+      AGInfo& info = AGInfo::Get(ograd_entries[i].node);
+      if (!idx.exist(ograd_entries[i].node.get())) continue;
+      size_t eid = idx.entry_id(ograd_entries[i]);
+      buff[eid] = info.outputs[0];
+      buff[eid].entry_ = ograd_entries[i];
+    }
+  } else {
+    states.reserve(num_forward_nodes);
+    for (size_t i = 0; i < num_forward_nodes; ++i) {
+      const AGInfo& info = dmlc::get<AGInfo>(idx[i].source->info);
+      states.emplace_back(info.state);
+      for (size_t j = 0; j < info.outputs.size(); ++j) {
+        size_t eid = idx.entry_id(i, j);
+        arrays[eid] = const_cast<NDArray*>(&(info.outputs[j]));
+        if (retain_graph || info.grad_req != kNullOp) ref_count[eid] = 1;
+      }
+    }
+    for (size_t i = 0; i < ograd_entries.size(); ++i) {
+      if (!idx.exist(ograd_entries[i].node.get())) continue;
+      AGInfo& info = AGInfo::Get(ograd_entries[i].node);
+      arrays[idx.entry_id(ograd_entries[i])] = &info.outputs[0];
+    }
+  }
+  for (size_t i = num_forward_outputs; i < graph.outputs.size(); ++i) {
+    size_t eid = idx.entry_id(graph.outputs[i]);
+    arrays[eid] = x_grads[i - num_forward_outputs];
+    ref_count[eid] = 1;
+  }
+
+  // Assign context
+  Context default_ctx = outputs[0]->ctx();
+
+  // Infer shape type
+  {
+    std::pair<uint32_t, uint32_t> node_range, entry_range;
+    node_range = {num_forward_nodes, idx.num_nodes()};
+    entry_range = {num_forward_entries, idx.num_node_entries()};
+
+    ShapeVector shapes;
+    shapes.reserve(idx.num_node_entries());
+    for (const auto& i : arrays) shapes.emplace_back(i->shape());
+    CheckAndInferShape(&graph, std::move(shapes), false,
+                       node_range, entry_range);
+
+    DTypeVector dtypes;
+    dtypes.reserve(idx.num_node_entries());
+    for (const auto& i : arrays) dtypes.emplace_back(i->dtype());
+    CheckAndInferType(&graph, std::move(dtypes), false,
+                      node_range, entry_range);
+
+    StorageTypeVector stypes;
+    stypes.reserve(idx.num_node_entries());
+    for (const auto& i : arrays) stypes.emplace_back(i->storage_type());
+    CheckAndInferStorageType(
+        &graph, default_ctx, std::move(stypes), false,
+        node_range, entry_range);
+  }
+
+  // Calculate ref count
+  for (size_t i = num_forward_nodes; i < idx.num_nodes(); ++i) {
+    for (const auto& j : idx[i].inputs) {
+       ++ref_count[idx.entry_id(j)];
+    }
+  }
+
+  // Assign reqs
+  std::vector<OpReqType> array_reqs(arrays.size(), kWriteTo);
+  for (size_t i = num_forward_entries; i < idx.num_node_entries(); ++i) {
+    if (ref_count[i] == 0) array_reqs[i] = kNullOp;
+  }
+  for (size_t i = num_forward_outputs; i < idx.outputs().size(); ++i) {
+    size_t eid = idx.entry_id(idx.outputs()[i]);
+    array_reqs[eid] = x_reqs[i - num_forward_outputs];
+  }
+
+  const auto& shapes = graph.GetAttr<ShapeVector>("shape");
+  const auto& dtypes = graph.GetAttr<DTypeVector>("dtype");
+  const auto& stypes = graph.GetAttr<StorageTypeVector>("storage_type");
+  for (size_t i = num_forward_entries; i < arrays.size(); ++i) {
+    if (!arrays[i]->is_none()) continue;
+    if (stypes[i] == kDefaultStorage) {
+      *arrays[i] = NDArray(shapes[i], default_ctx, true, dtypes[i]);
+    } else {
+      *arrays[i] = NDArray(static_cast<NDArrayStorageType>(stypes[i]),
+                           shapes[i], default_ctx, true, dtypes[i]);
+    }
+  }
+
+  // Execution
+
+  bool prev_recording = set_is_recording(create_graph);
+  bool prev_training = set_is_training(is_train);
+
+  RunGraph(retain_graph, default_ctx, idx, arrays, num_forward_nodes, idx.num_nodes(),
+           std::move(array_reqs), std::move(ref_count), &states);
+
+  set_is_recording(prev_recording);
+  set_is_training(prev_training);
+
+  // Clear history
+  if (!retain_graph) {
+    nnvm::DFSVisit(sym.outputs, [&](const nnvm::NodePtr& n) {
+      AGInfo::Clear(n);
+      n->inputs.clear();
+    });
+  }
+
+  if (variables.size()) {
+    return x_grads;
+  }
+  return {};
+}
+
+}  // namespace mxnet
